[RENAME] : find 파일이동 및 Page와 components 병합
