+++ src/pages/FindId/Page.module.css
@@ -1,17 +1,16 @@
 .container {
   display: flex;
-  flex-direction: column;
-  align-items: center;
   justify-content: center;
   background: var(--secondary-g);
+  padding: 4rem;
 }
 
 .title {
-  text-align: center;
   font-size: var(--large-2);
   font-weight: bold;
   margin-bottom: var(--space-5);
 }
+
 .findIdForm {
   display: flex;
   flex-direction: column;
@@ -21,16 +20,22 @@
   height: 30rem;
   padding: 7.125rem;
   border-radius: 2rem;
-  background: var(--light-gray, linear-gradient(0deg, #e2e2e2 0%, #fff 100%));
+  background: var(--gray-g);
+}
+
+.inputForm {
+  max-width: 25rem;
+  margin-bottom: 0.625rem;
+  border: 1px solid var(--white);
+  border-radius: 0.625rem;
+  font-size: var(--medium);
 }
 .submitButton {
   width: 20rem;
   height: 5rem;
   font-size: var(--large);
-  font-weight: 400;
   border: 1px solid var(--black);
   border-radius: var(--radius-xl);
   cursor: pointer;
-  align-self: center;
   text-align: center;
 }
